+++ src/components/TableWithInlineEditing.js
@@ -196,6 +196,8 @@ const Row = ({
                                                     onChange={handleCellChange}
                                                     items={column.options}
                                                     value={rowContent[column.key]}
+                                                    disabled={column.notEditable}
+                                                    required={column.required}
                                                     propertyName={column.key}
                                                     label=""
                                                     allowNoValue={false}
@@ -204,6 +206,8 @@ const Row = ({
                                                 <InputField
                                                     type={column.type}
                                                     value={rowContent[column.key]}
+                                                    disabled={column.notEditable}
+                                                    required={column.required}
                                                     onChange={handleCellChange}
                                                     propertyName={column.key}
                                                 />
@@ -255,7 +259,8 @@ Row.propTypes = {
         PropTypes.shape({
             title: PropTypes.string,
             type: PropTypes.string,
-            key: PropTypes.oneOfType([PropTypes.string, PropTypes.number])
+            key: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+            notEditable: PropTypes.bool
         })
     ).isRequired,
     removeRow: PropTypes.func.isRequired,